requirements.txt
requests
python-dotenv
scikit-learn
numpy
chromadb
rank-bm25
nltk
pandas
langchain-groq
langchain-core
google-generativeai
langchain
rapidfuzz
langgraph
fastapi
uvicorn[standard]
pydantic
